requirements.txt
# 核心依赖
pyautocad==0.2.0          # AutoCAD 接口库
pywin32==308              # Windows COM 支持
#pythoncom         # COM 接口支持

# 配置文件处理
PyYAML==6.0.1             # YAML 配置文件解析

# 数据处理与工具
numpy==1.24.3             # 矩阵计算与坐标生成
dataclasses==0.6          # Python 3.6 兼容 dataclass（Python 3.7+ 可移除）

# 命令行工具
argparse==1.4.0            # 命令行参数解析（Python 标准库，无需安装）

# 开发工具（可选）
black==23.3.0             # 代码格式化工具
pylint==2.17.4            # 代码静态检查